\name{ComtradeAPI-package}
\alias{ComtradeAPI-package}
\alias{ComtradeAPI}
\docType{package}
\title{
\packageTitle{ComtradeAPI}
}
\description{
\packageDescription{ComtradeAPI}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ComtradeAPI}
\packageIndices{ComtradeAPI}
Use this package to downlaod trade data through the comtrade API
}
\author{
\packageAuthor{ComtradeAPI}

Maintainer: \packageMaintainer{ComtradeAPI}
}
\references{
See [comtrade](https://comtradedeveloper.un.org) for more information
}