% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBulkData.R
\name{getBulkData}
\alias{getBulkData}
\title{Download bulk data from Comtrade}
\usage{
getBulkData(
  base_url = "https://comtradeapi.un.org/bulk/v1/get",
  key = key,
  fileLocation,
  typeCode = "C",
  freqCode = "A",
  clCode = "HS",
  reporterCode = NULL,
  period = NULL,
  publishedDateFrom = NULL,
  publishedDateTo = NULL
)
}
\arguments{
\item{key}{API key for the comtrade API (MANDATORY)}

\item{typeCode}{Type of trade: C for commodities and S for service (MANDATORY)}

\item{freqCode}{Trade frequency: A for annual and M for monthly (MANDATORY)}

\item{clCode}{Trade (IMTS) classification: HS, SITC, BEC or EBOPS (MANDATORY)}

\item{reporterCode}{Reporter Code (possible values are M49 code of the countries)
See \code{ComtradeAPI::reporterCode} for country, code pairs (NOT MANDATORY)}

\item{period}{Year or month. Year should be 4 digit year. (NOT MANDATORY)
Month should be six digit integer with the values of the form YYYYMM. Ex: 201002 for 2010 February.
Multi value input should be in the form of csv (Codes separated by comma (,))}

\item{publishedDateFrom}{Publication date From YYYY-MM-DD (NOT MANDATORY)}

\item{publishedDateTo}{Publication date To YYYY-MM-DD (NOT MANDATORY)}
}
\description{
getBulkData(
 url,
 key,
 typeCode = "C",
 freqCode = "A",
 clCode = "HS",
 reporterCode = NULL,
 period = NULL,
 publishedDateFrom = NULL,
 publishedDateTo = NULL,
 fileLocation = NULL
)
}
