% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Download data from comtrade}
\usage{
getData(
  base_url = "https://comtradeapi.un.org/data/v1/get/",
  key = getPrimaryKey(),
  typeCode = "C",
  freqCode = "A",
  clCode = "HS",
  reporterCode = NULL,
  period = NULL,
  partnerCode = NULL,
  partner2Code = NULL,
  cmdCode = NULL,
  flowCode = NULL,
  customsCode = NULL,
  motCode = NULL,
  aggregateBy = NULL,
  breakdownMode = NULL,
  includeDesc = NULL,
  maxRecords = NULL
)
}
\arguments{
\item{key}{API key for the comtrade API (MANDATORY)}

\item{typeCode}{Type of trade: C for commodities and S for service (MANDATORY)}

\item{freqCode}{Trade frequency: A for annual and M for monthly (MANDATORY)}

\item{clCode}{Trade (IMTS) classification: HS, SITC, BEC or EBOPS (MANDATORY)
Available Options: HS, H0 - H6, S1-S4, BE, BE5, Eb, EB10, EB02, EBSDMX}

\item{reporterCode}{Reporter Code (possible values are M49 code of the countries)
See \code{ComtradeAPI::reporterCode} for country, code pairs (NOT MANDATORY)}

\item{period}{Year or month. Year should be 4 digit year. (NOT MANDATORY)
Month should be six digit integer with the values of the form YYYYMM. Ex: 201002 for 2010 February.
Multi value input should be in the form of csv (Codes separated by comma (,))}

\item{partnerCode}{Partner code (Possible values are M49 code of the countries separated by comma (,))}

\item{cmdCode}{Commodity code. Multi value input should be in the form of csv (Codes separated by comma (,))}

\item{flowCode}{Trade flow code. Multi value input should be in the form of csv (Codes separated by comma (,))}

\item{customsCode}{Customs code. Multi value input should be in the form of csv (Codes separated by comma (,))}

\item{motCode}{Mode of transport code. Multi value input should be in the form of csv (Codes separated by comma (,))}

\item{aggregateBy}{Add parameters in csv list on which you want the results to be aggregated}

\item{breakdownMode}{Mode to choose from}

\item{includeDesc}{Include descriptions of data variables}

\item{maxRecords}{Maximum number of rows the query should return}

\item{Second}{partner/consignment code (Possible values are M49 code of the countries separated by comma (,))}
}
\description{
Download data from comtrade
}
