% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{setPrimaryKey}
\alias{setPrimaryKey}
\alias{getPrimaryKey}
\title{set and retreive comtrade API key}
\usage{
setPrimaryKey(key = NULL)

getPrimaryKey()
}
\arguments{
\item{key}{your comtrade API key}
}
\description{
Use thesee functions to supply your comtrade API key which is needed for the other functions
in this package to interact with the API
}
\section{Functions}{
\itemize{
\item \code{setPrimaryKey()}: set API key

\item \code{getPrimaryKey()}: get API key

}}
