% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/previewBulkData.R
\name{previewBulkData}
\alias{previewBulkData}
\title{Preview files to be downloaded from comtradeAPI::getBulkData()}
\usage{
previewBulkData(
  base_url = "https://comtradeapi.un.org/bulk/v1/get",
  key = getPrimaryKey(),
  typeCode = "C",
  freqCode = "A",
  clCode = "HS",
  reporterCode = NULL,
  period = NULL,
  publishedDateFrom = NULL,
  publishedDateTo = NULL
)
}
\arguments{
\item{key}{API key for the comtrade API (MANDATORY)}

\item{typeCode}{Type of trade: C for commodities and S for service (MANDATORY)}

\item{freqCode}{Trade frequency: A for annual and M for monthly (MANDATORY)}

\item{clCode}{Trade (IMTS) classification: HS, SITC, BEC or EBOPS (MANDATORY)}

\item{reporterCode}{Reporter Code (possible values are M49 code of the countries)
See \code{ComtradeAPI::reporterCode} for country, code pairs (NOT MANDATORY)}

\item{period}{Year or month. Year should be 4 digit year. (NOT MANDATORY)
Month should be six digit integer with the values of the form YYYYMM. Ex: 201002 for 2010 February.
Multi value input should be in the form of csv (Codes separated by comma (,))}

\item{publishedDateFrom}{Publication date From YYYY-MM-DD (NOT MANDATORY)}

\item{publishedDateTo}{Publication date To YYYY-MM-DD (NOT MANDATORY)}
}
\description{
previewBulkData(
 base_url = "https://comtradeapi.un.org/bulk/v1/get",
 key = getPrimaryKey(),
 typeCode = "C",
 freqCode = "A",
 clCode = "HS",
 reporterCode = NULL,
 period = NULL,
 publishedDateFrom = NULL,
 publishedDateTo = NULL
)
}
\details{
This function allows you to catch a glimpse at the links which would be downloaded
were you to pass the same parameters in \code{comtradeAPI::getBulkData}
This includes information such as the size of each file as well as its contents

When you wish to download multiple files in bulk, it is recommended to run this function first and store the output as a dataframe
to check which files you will be downloading. In case you do not wish to download all files, simply remove
the rows which you do not want to download and then pass the dataframe to the parameter 'data' in \code{comtradeAPI::getBulkDataFromLink}
}
