\name{comtradeAPI-package}
\alias{comtradeAPI-package}
\alias{comtradeAPI}
\docType{package}
\title{
\packageTitle{comtradeAPI}
}
\description{
\packageDescription{comtradeAPI}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{comtradeAPI}
\packageIndices{comtradeAPI}
Use this package to downlaod trade data through the comtrade API
}
\author{
\packageAuthor{comtradeAPI}

Maintainer: \packageMaintainer{comtradeAPI}
}
\references{
See [comtrade](https://comtradedeveloper.un.org) for more information
}